Rev up vr_hwc to composer@2.3
Fix for issue: "[GSI XR] VrFlinger doesn't work on GSI/MTP845.

Adding a new IVRComposerClient deriving from the hardware graphics
composer 2.3. This is needed to be able to return the edid data in the
function getDisplayIdentificationData().

Bug: 137325030
Bug: 138938154
Bug: 137448042

Test: Built the firmware and ran the dvr_display-test, which passes
now. There are multiple changes that are required to get all of this
working.

Change-Id: Id413413544f25bfd800df3a25dcf05ef934134fb