Revert "Remove vendor interface stability from internal power policy interface"
This reverts commit 37c83d8a74c1641f1efb8174e70febe2ac1ac686.

The original commit removed vintf stability from the tip-of-tree version
of the interface, leaving the frozen version 1 untouched. However, the
removal of the `stability: vintf` from the Android.bp file is causing a
problem because the flag is still needed for the AIDL compiler to parse
the frozen AIDL file at version 1.

Just adding `stability: vintf` to the build file is not enough because
then the tip-of-tree version reports the opposite issue; "... does not
have VINTF stability, but this interface requires it."

Since the removal of vintf stability seems to have been nice-to-have,
reverting the original commit, to unblock the TM-dev AIDL API freeze.

Bug: 214428116
Bug: 225941299
Test: m
Change-Id: I009c21cc521544fdadf5e99b637745ecaa45208c