cameraservice: Add Surface to OutputConfiguration
The current interface uses NativeHandle to parcel NDK Surface
over the stable interface. However, parceling objects as
NativeHandle is deprecated as it was dependent on HIDL Token
Manager, and interfaces must use Surfaces instead which are now
parcelable over stable interface.

This CL updates OutputConfiguration to add a `surfaces` field
which should be used from here on out. The existing
windowHandles field is marked deprecated.

Bug: 283283111
Test: Camera vNDK implementation updated to use the
      new field with updated tests. FaceHAL also now
      uses the new field.
Change-Id: Ia8d428e940689397acab9cb9b494132eb0a5ecdc